practice/data-pipeline-project/requirements.txt
aiohttp                   3.8.5
aiosignal                 1.3.1
annotated-types           0.5.0
appnope                   0.1.3
astroid                   2.15.6
asttokens                 2.2.1
async-timeout             4.0.2
attrs                     23.1.0
backcall                  0.2.0
black                     23.7.0
certifi                   2023.7.22
charset-normalizer        3.2.0
click                     8.1.6
decorator                 5.1.1
dill                      0.3.7
duckdb                    0.8.0
executing                 1.2.0
frozenlist                1.4.0
idna                      3.4
ipython                   8.14.0
isort                     5.12.0
jedi                      0.18.2
jsonschema                4.18.4
jsonschema-specifications 2023.7.1
lazy-object-proxy         1.9.0
matplotlib-inline         0.1.6
mccabe                    0.7.0
multidict                 6.0.4
mypy                      1.4.1
mypy-extensions           1.0.0
numpy                     1.25.1
packaging                 23.1
pandas                    2.0.3
parso                     0.8.3
pathspec                  0.11.2
pexpect                   4.8.0
pickleshare               0.7.5
pip                       23.2.1
platformdirs              3.10.0
prompt-toolkit            3.0.39
ptyprocess                0.7.0
pure-eval                 0.2.2
pydantic                  2.0.3
pydantic_core             2.3.0
Pygments                  2.15.1
pylint                    2.17.5
python-dateutil           2.8.2
python-dotenv             1.0.0
pytz                      2023.3
referencing               0.30.0
requests                  2.31.0
rpds-py                   0.9.2
setuptools                65.5.0
six                       1.16.0
stack-data                0.6.2
tomli                     2.0.1
tomlkit                   0.12.1
traitlets                 5.9.0
typing_extensions         4.7.1
tzdata                    2023.3
urllib3                   2.0.4
wcwidth                   0.2.6
wrapt                     1.15.0
yarl                      1.9.2